[Refactor] findMeetingWithMemberCount 메서드 member 패키지로 이동 (#46)
