+++ src/main/kotlin/com/ringgo/domain/member/repository/MemberRepositoryCustom.kt
@@ -1,5 +1,8 @@
 package com.ringgo.domain.member.repository
 
-interface MemberRepositoryCustom {
+import com.ringgo.domain.meeting.dto.MeetingWithMemberCount
+import java.util.*
 
+interface MemberRepositoryCustom {
+    fun findMeetingWithMemberCount(userId: UUID): List<MeetingWithMemberCount>
 }
